move share to component
