+++ frontend/src/features/collections/index.ts
@@ -3,3 +3,4 @@ import("./collection-items-dialog");
 import("./collection-metadata-dialog");
 import("./collection-workflow-list");
 import("./select-collection-access");
+import("./share-collection");
